refactor: revert entities & add phase data
