+++ src/data/phase/beltaine.tsx
@@ -1,9 +1,10 @@
-import type { Phase } from './types'
+import type { PhaseDTO } from './dto.type'
 
-export const beltaine: Phase = {
+export const beltaine: PhaseDTO = {
   number: 3,
-  name: 'beltaine',
-  description:
+  phaseName: 'beltaine',
+  phaseDescription:
     'The third phase is all about token dynamics! Druids will engage in various node and community tasks with their precious tokens. Challenges will include some IBC-related tasks to open Nemeton to the interchain world...ime. Maintenance tasks and upgrades will be performed to test different kinds of state migrations.',
-  status: 'coming'
+  status: 'coming',
+  tasks: []
 }
