--- src/entity/phases/index.ts
@@ -1,7 +1,0 @@
-import { sidh } from './sidh'
-import { imbolc } from './imbolc'
-import { beltaine } from './beltaine'
-
-const Phases = { imbolc, beltaine, sidh }
-
-export default Phases
